+++ src/components/Toast/index.ts
@@ -1,6 +1,7 @@
 export { Toast } from "./Toast";
 export { useToast } from "./Toast.context";
 export type {
+  ToastAction,
   ToastData,
   ToastVariant,
   ToastPosition,